+++ CHANGELOG.md
@@ -38,6 +38,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop the checks for the zip_archive PHP extension (#175)
 
 ### Fixed
+- Extract the imported ZIPs within typo3temp, not fileadmin (#217)
 - Set the sorting when adding attachments to realty objects (#212)
 - Keep the sorting when converting images and documents to FAL (#211)
 - Use the correct path name when PHP-linting the tests (#192)